[CHORE] 롤백 후 extension 적용
